+++ src/main/java/net/geforcemods/securitycraft/blockentities/ReinforcedDispenserBlockEntity.java
@@ -1,0 +1,160 @@
+package net.geforcemods.securitycraft.blockentities;
+
+import java.util.EnumMap;
+import java.util.Map;
+
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.api.Owner;
+import net.geforcemods.securitycraft.inventory.InsertOnlyInvWrapper;
+import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.util.BlockUtils;
+import net.minecraft.block.BlockState;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.network.NetworkManager;
+import net.minecraft.network.play.server.SUpdateTileEntityPacket;
+import net.minecraft.tileentity.DispenserTileEntity;
+import net.minecraft.tileentity.TileEntityType;
+import net.minecraft.util.Direction;
+import net.minecraft.util.NonNullList;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.util.LazyOptional;
+import net.minecraftforge.items.CapabilityItemHandler;
+import net.minecraftforge.items.IItemHandler;
+
+public class ReinforcedDispenserBlockEntity extends DispenserTileEntity implements IOwnable, IModuleInventory {
+	private LazyOptional<IItemHandler> insertOnlyHandler;
+	private NonNullList<ItemStack> modules = NonNullList.withSize(getMaxNumberOfModules(), ItemStack.EMPTY);
+	private Owner owner = new Owner();
+	private Map<ModuleType, Boolean> moduleStates = new EnumMap<>(ModuleType.class);
+
+	public ReinforcedDispenserBlockEntity(TileEntityType<?> type) {
+		super(type);
+	}
+
+	public ReinforcedDispenserBlockEntity() {
+		super(SCContent.REINFORCED_DISPENSER_BLOCK_ENTITY.get());
+	}
+
+	@Override
+	protected ITextComponent getDefaultName() {
+		return new TranslationTextComponent("block.securitycraft.reinforced_dispenser");
+	}
+
+	@Override
+	public void load(BlockState state, CompoundNBT tag) {
+		super.load(state, tag);
+
+		owner.load(tag);
+		modules = readModuleInventory(tag);
+		moduleStates = readModuleStates(tag);
+	}
+
+	@Override
+	public CompoundNBT save(CompoundNBT tag) {
+		super.save(tag);
+
+		if (owner != null)
+			owner.save(tag, needsValidation());
+
+		writeModuleInventory(tag);
+		writeModuleStates(tag);
+		return tag;
+	}
+
+	@Override
+	public CompoundNBT getUpdateTag() {
+		return save(new CompoundNBT());
+	}
+
+	@Override
+	public SUpdateTileEntityPacket getUpdatePacket() {
+		return new SUpdateTileEntityPacket(worldPosition, 1, getUpdateTag());
+	}
+
+	@Override
+	public void onDataPacket(NetworkManager net, SUpdateTileEntityPacket packet) {
+		load(getBlockState(), packet.getTag());
+	}
+
+	@Override
+	public Owner getOwner() {
+		return owner;
+	}
+
+	@Override
+	public void setOwner(String uuid, String name) {
+		owner.set(uuid, name);
+		setChanged();
+	}
+
+	@Override
+	public boolean enableHack() {
+		return true;
+	}
+
+	@Override
+	public ItemStack getStackInSlot(int slot) {
+		return slot >= 100 ? getModuleInSlot(slot) : super.getItem(slot);
+	}
+
+	@Override
+	public ItemStack getItem(int slot) {
+		return getStackInSlot(slot);
+	}
+
+	@Override
+	public <T> LazyOptional<T> getCapability(Capability<T> cap, Direction side) {
+		if (cap == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)
+			return BlockUtils.isAllowedToExtractFromProtectedBlock(side, this) ? super.getCapability(cap, side) : getInsertOnlyHandler().cast();
+		else
+			return super.getCapability(cap, side);
+	}
+
+	@Override
+	public void invalidateCaps() {
+		if (insertOnlyHandler != null)
+			insertOnlyHandler.invalidate();
+
+		super.invalidateCaps();
+	}
+
+	@Override
+	public void reviveCaps() {
+		insertOnlyHandler = null; //recreated in getInsertOnlyHandler
+		super.reviveCaps();
+	}
+
+	private LazyOptional<IItemHandler> getInsertOnlyHandler() {
+		if (insertOnlyHandler == null)
+			insertOnlyHandler = LazyOptional.of(() -> new InsertOnlyInvWrapper(this));
+
+		return insertOnlyHandler;
+	}
+
+	@Override
+	public ModuleType[] acceptedModules() {
+		return new ModuleType[] {
+				ModuleType.ALLOWLIST
+		};
+	}
+
+	@Override
+	public NonNullList<ItemStack> getInventory() {
+		return modules;
+	}
+
+	@Override
+	public boolean isModuleEnabled(ModuleType module) {
+		return hasModule(module) && moduleStates.get(module) == Boolean.TRUE; //prevent NPE
+	}
+
+	@Override
+	public void toggleModuleState(ModuleType module, boolean shouldBeEnabled) {
+		moduleStates.put(module, shouldBeEnabled);
+	}
+}