+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedDropperBlock.java
@@ -1,0 +1,77 @@
+package net.geforcemods.securitycraft.blocks.reinforced;
+
+import net.geforcemods.securitycraft.blockentities.ReinforcedDropperBlockEntity;
+import net.minecraft.block.Block;
+import net.minecraft.block.Blocks;
+import net.minecraft.dispenser.DefaultDispenseItemBehavior;
+import net.minecraft.dispenser.IDispenseItemBehavior;
+import net.minecraft.dispenser.ProxyBlockSource;
+import net.minecraft.inventory.IInventory;
+import net.minecraft.item.ItemStack;
+import net.minecraft.tileentity.HopperTileEntity;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.Direction;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockReader;
+import net.minecraft.world.server.ServerWorld;
+import net.minecraftforge.common.util.Constants;
+import net.minecraftforge.items.VanillaInventoryCodeHooks;
+
+public class ReinforcedDropperBlock extends ReinforcedDispenserBlock {
+	private static final IDispenseItemBehavior DISPENSE_BEHAVIOUR = new DefaultDispenseItemBehavior();
+
+	public ReinforcedDropperBlock(Properties properties) {
+		super(properties);
+	}
+
+	@Override
+	public IDispenseItemBehavior getDispenseMethod(ItemStack stack) {
+		return DISPENSE_BEHAVIOUR;
+	}
+
+	@Override
+	public TileEntity newBlockEntity(IBlockReader level) {
+		return new ReinforcedDropperBlockEntity();
+	}
+
+	@Override
+	protected void dispenseFrom(ServerWorld level, BlockPos pos) {
+		if (level.getBlockEntity(pos) instanceof ReinforcedDropperBlockEntity) {
+			ProxyBlockSource source = new ProxyBlockSource(level, pos);
+			ReinforcedDropperBlockEntity be = source.getEntity();
+			int randomSlot = be.getRandomSlot();
+
+			if (randomSlot < 0)
+				level.levelEvent(Constants.WorldEvents.DISPENSER_FAIL_SOUND, pos, 0);
+			else {
+				ItemStack dispenseStack = be.getItem(randomSlot);
+
+				if (!dispenseStack.isEmpty() && VanillaInventoryCodeHooks.dropperInsertHook(level, pos, be, randomSlot, dispenseStack)) {
+					Direction direction = level.getBlockState(pos).getValue(FACING);
+					IInventory inventory = HopperTileEntity.getContainerAt(level, pos.relative(direction));
+					ItemStack afterDispenseStack;
+
+					if (inventory == null)
+						afterDispenseStack = DISPENSE_BEHAVIOUR.dispense(source, dispenseStack);
+					else {
+						afterDispenseStack = HopperTileEntity.addItem(be, inventory, dispenseStack.copy().split(1), direction.getOpposite());
+
+						if (afterDispenseStack.isEmpty()) {
+							afterDispenseStack = dispenseStack.copy();
+							afterDispenseStack.shrink(1);
+						}
+						else
+							afterDispenseStack = dispenseStack.copy();
+					}
+
+					be.setItem(randomSlot, afterDispenseStack);
+				}
+			}
+		}
+	}
+
+	@Override
+	public Block getVanillaBlock() {
+		return Blocks.DROPPER;
+	}
+}