+++ Changelog.md
@@ -4,6 +4,7 @@
 - New: The Sentry Remote Access tool now displays the last known name of a Sentry if the Sentry is out of range
 - New: The Portable Radar now always shows its coordinates in messages, even when it has a custom name
 - New: Crafting Recipes for Reinforced Pistons and Reinforced Sticky Pistons
+- New: Reinforced Dispenser and Reinforced Dropper
 - Change: Sentries now shoot half as frequently to match the shooting rate in newer Minecraft versions
 - Change: The Reinforced Hopper screen now shows "Reinforced Hopper" instead of "Item Hopper" as its default inventory title
 - Change: The hitbox of the Sentry has been changed, which means that only the base of Sentries can be interacted with now